% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-matrix.R
\name{as_tibble.gs_matrix}
\alias{as_tibble.gs_matrix}
\title{Convert a \code{gs_matrix} to long form}
\usage{
\method{as_tibble}{gs_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{gs_matrix}.}

\item{...}{Ignored.}
}
\value{
A tibble with \code{pathway_id}, \code{pathway_name}, \code{sample}, \code{score},
\code{database}, \code{method}, \code{score_type}, plus any \code{sample_data} columns.
}
\description{
Convert a \code{gs_matrix} to long form
}
