% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-result.R
\name{as_tibble.gs_result}
\alias{as_tibble.gs_result}
\title{Drop the \code{gs_result} class}
\usage{
\method{as_tibble}{gs_result}(x, ...)
}
\arguments{
\item{x}{A \code{gs_result}.}

\item{...}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.}
}
\value{
A plain tibble.
}
\description{
Drop the \code{gs_result} class
}
