% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkiRNA-package.R
\docType{package}
\name{bulkiRNA-package}
\alias{bulkiRNA}
\alias{bulkiRNA-package}
\title{bulkiRNA: bulk RNA-seq differential expression and gene-set analysis}
\description{
Four layers, one direction of flow:
}
\details{
\preformatted{
  gsdb_*        gs_test() / gs_score()     gs_result / gs_matrix     gs_plot_*
  providers --> compute               -->  data objects         -->  renderers
}

Compute functions return objects and never plot or write; \code{gs_write()} /
\code{gs_read()} / \code{gs_save()} are the only disk access; \verb{gs_plot_*} take an
object and return a ggplot.

See \link[=gs_result-class]{gs_result} and \link[=gs_matrix-class]{gs_matrix} for the shared data contracts.
}
\section{Reference data}{

Bundled processed gene-set databases live in \code{inst/extdata} and resolve via
\code{system.file("extdata", ..., package = "bulkiRNA")}. Raw source files are in
the git checkout only and are not shipped.
}

\author{
\strong{Maintainer}: Anton Zhelonkin \email{antonz@uchicago.edu}

Authors:
\itemize{
  \item Anton Zhelonkin \email{antonz@uchicago.edu}
}

}
\keyword{internal}
