% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{.BULKI_PALETTES}
\alias{.BULKI_PALETTES}
\title{The vendored palette table}
\usage{
.BULKI_PALETTES
}
\value{
A named list of character vectors of hex colours.
}
\description{
The vendored palette table
}
\keyword{internal}
