% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{.BULKI_QUALITATIVE}
\alias{.BULKI_QUALITATIVE}
\title{The default qualitative palette name}
\usage{
.BULKI_QUALITATIVE
}
\value{
A character scalar.
}
\description{
The default qualitative palette name
}
\keyword{internal}
