% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref-path.R
\name{.clear_ref_resolutions}
\alias{.clear_ref_resolutions}
\title{Clear the session reference-data record}
\usage{
.clear_ref_resolutions()
}
\value{
\code{NULL}, invisibly.
}
\description{
Clear the session reference-data record
}
\keyword{internal}
