% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresh.R
\name{.coresh_empty_search}
\alias{.coresh_empty_search}
\title{Empty CoReSh search result}
\usage{
.coresh_empty_search()
}
\value{
A zero-row tibble with the search schema.
}
\description{
Empty CoReSh search result
}
\keyword{internal}
