% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresh.R
\name{.coresh_read_chunk}
\alias{.coresh_read_chunk}
\title{Read one CoReSh chunk}
\usage{
.coresh_read_chunk(path)
}
\arguments{
\item{path}{Path to a qs2 chunk.}
}
\value{
The deserialized list of dataset objects.
}
\description{
Read one CoReSh chunk
}
\keyword{internal}
