% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresh.R
\name{.coresh_validate_object}
\alias{.coresh_validate_object}
\title{Validate a CoReSh dataset object}
\usage{
.coresh_validate_object(obj, context = "`obj`")
}
\arguments{
\item{obj}{A dataset object from a CoReSh chunk.}

\item{context}{Description used in an error message.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate a CoReSh dataset object
}
\keyword{internal}
