% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresh-sets.R
\name{.coresh_validate_set_inputs}
\alias{.coresh_validate_set_inputs}
\title{Validate inputs for CoReSh set construction}
\usage{
.coresh_validate_set_inputs(top_hits, queries)
}
\arguments{
\item{top_hits}{A data frame of ranked CoReSh hits.}

\item{queries}{Named list of Entrez queries.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate inputs for CoReSh set construction
}
\keyword{internal}
