% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_check_heights}
\alias{.grs_check_heights}
\title{Validate \code{panel_heights}}
\usage{
.grs_check_heights(panel_heights)
}
\arguments{
\item{panel_heights}{The renderer's \code{panel_heights}.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate \code{panel_heights}
}
\keyword{internal}
