% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_check_ylim}
\alias{.grs_check_ylim}
\title{Validate \code{es_ylim}}
\usage{
.grs_check_ylim(es_ylim)
}
\arguments{
\item{es_ylim}{The renderer's \code{es_ylim}, possibly \code{NULL}.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate \code{es_ylim}
}
\keyword{internal}
