% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_compose}
\alias{.grs_compose}
\title{Compose the three panels}
\usage{
.grs_compose(panels, panel_heights, legend_position)
}
\arguments{
\item{panels}{List of three \code{ggplot}s, top to bottom.}

\item{panel_heights}{Length-3 numeric height ratio.}

\item{legend_position}{Legend placement.}
}
\value{
A \code{patchwork}.
}
\description{
Compose the three panels
}
\keyword{internal}
