% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_hide_x}
\alias{.grs_hide_x}
\title{Hide the x axis on a panel that is not the bottom one}
\usage{
.grs_hide_x()
}
\value{
A ggplot2 theme.
}
\description{
Hide the x axis on a panel that is not the bottom one
}
\keyword{internal}
