% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_is_sets}
\alias{.grs_is_sets}
\title{Is \code{x} shaped like a named list of gene sets?}
\usage{
.grs_is_sets(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is \code{x} shaped like a named list of gene sets?
}
\keyword{internal}
