% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_legend_theme}
\alias{.grs_legend_theme}
\title{Legend placement}
\usage{
.grs_legend_theme(legend_position, legend_pos)
}
\arguments{
\item{legend_position}{One of \code{"inside"}, \code{"right"}, \code{"bottom"}, \code{"none"}.}

\item{legend_pos}{Length-2 numeric inside-legend position.}
}
\value{
A ggplot2 theme.
}
\description{
Legend placement
}
\keyword{internal}
