% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.GRS_PANEL_HEIGHTS}
\alias{.GRS_PANEL_HEIGHTS}
\title{Default ES : ticks : metric height ratio}
\usage{
.GRS_PANEL_HEIGHTS
}
\value{
A length-3 numeric vector.
}
\description{
The single source of truth for panel proportions.
}
\keyword{internal}
