% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-running.R
\name{.grs_sets}
\alias{.grs_sets}
\title{Resolve gene-set membership for \code{\link[=gs_plot_running]{gs_plot_running()}}}
\usage{
.grs_sets(x, db)
}
\arguments{
\item{x}{The renderer's \code{x}.}

\item{db}{The renderer's \code{db}, possibly \code{NULL}.}
}
\value{
A named list of character vectors.
}
\description{
Resolve gene-set membership for \code{\link[=gs_plot_running]{gs_plot_running()}}
}
\keyword{internal}
