% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-matrix.R
\name{.gs_align_sample_data}
\alias{.gs_align_sample_data}
\title{Align sample metadata to matrix columns}
\usage{
.gs_align_sample_data(sample_data, samples)
}
\arguments{
\item{sample_data}{Data frame of sample metadata.}

\item{samples}{Character vector of column names to align to.}
}
\value{
\code{sample_data} reordered to \code{samples}.
}
\description{
Align sample metadata to matrix columns
}
\keyword{internal}
