% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-dot.R
\name{.gs_check_compare}
\alias{.gs_check_compare}
\title{Validate the \code{compare} argument}
\usage{
.gs_check_compare(compare)
}
\arguments{
\item{compare}{\code{NULL}, \code{"contrast"} or \code{"database"}.}
}
\value{
\code{compare}, unchanged.
}
\description{
Validate the \code{compare} argument
}
\keyword{internal}
