% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-ops.R
\name{.gs_check_result}
\alias{.gs_check_result}
\title{Error unless \code{x} is a \code{gs_result}}
\usage{
.gs_check_result(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{x}, invisibly.
}
\description{
Error unless \code{x} is a \code{gs_result}
}
\keyword{internal}
