% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-ranks.R
\name{.gs_collapse_ranks}
\alias{.gs_collapse_ranks}
\title{Collapse duplicated gene identifiers in a rank vector}
\usage{
.gs_collapse_ranks(v, how)
}
\arguments{
\item{v}{A named numeric vector.}

\item{how}{\code{"max_abs"} or \code{"mean"}.}
}
\value{
A named numeric vector with unique names.
}
\description{
Collapse duplicated gene identifiers in a rank vector
}
\keyword{internal}
