% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-coregulation.R
\name{.gs_coregulation_logical}
\alias{.gs_coregulation_logical}
\title{Validate a logical GESECA argument}
\usage{
.gs_coregulation_logical(x, name)
}
\arguments{
\item{x}{Value to validate.}

\item{name}{Argument name without backticks.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a logical GESECA argument
}
\keyword{internal}
