% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-coregulation.R
\name{.gs_coregulation_validate_expr}
\alias{.gs_coregulation_validate_expr}
\title{Validate a GESECA expression matrix}
\usage{
.gs_coregulation_validate_expr(expr)
}
\arguments{
\item{expr}{Object supplied as \code{expr} to \code{\link[=gs_coregulation]{gs_coregulation()}}.}
}
\value{
\code{expr}, invisibly.
}
\description{
Validate a GESECA expression matrix
}
\keyword{internal}
