% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-test.R
\name{.gs_db_sets}
\alias{.gs_db_sets}
\title{Strip the \code{gs_db} class down to the bare named list of character vectors}
\usage{
.gs_db_sets(db)
}
\arguments{
\item{db}{A \code{gs_db}.}
}
\value{
A named list of character vectors.
}
\description{
Strip the \code{gs_db} class down to the bare named list of character vectors
}
\keyword{internal}
