% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-plot-utils.R
\name{.gs_plot_check_result}
\alias{.gs_plot_check_result}
\title{Check that an object is a usable \code{gs_result}}
\usage{
.gs_plot_check_result(x, arg = "x")
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Name of the argument being checked, for the error message.}
}
\value{
\code{x}, invisibly.
}
\description{
Check that an object is a usable \code{gs_result}
}
\keyword{internal}
