% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-io.R
\name{.gs_warn_unmanifested}
\alias{.gs_warn_unmanifested}
\title{Warn about globbed files the last \code{gs_write()} did not produce}
\usage{
.gs_warn_unmanifested(dir, name, found)
}
\arguments{
\item{dir}{The directory being read.}

\item{name}{The file-name stem.}

\item{found}{Absolute paths the glob returned.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Warn about globbed files the last \code{gs_write()} did not produce
}
\keyword{internal}
