% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdb-rebuild.R
\name{.gsdb_human_to_mouse}
\alias{.gsdb_human_to_mouse}
\title{Map human gene symbols to mouse orthologs}
\usage{
.gsdb_human_to_mouse(db, verbose = FALSE)
}
\arguments{
\item{db}{A \code{\link[=gs_db]{gs_db()}} holding human symbols.}

\item{verbose}{Logical(1); message mapping statistics.}
}
\value{
A \code{gs_db} with mouse symbols and \code{species = "Mus musculus"}.
}
\description{
Map human gene symbols to mouse orthologs
}
\keyword{internal}
