% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdb-file.R
\name{.gsdb_parse_gmt}
\alias{.gsdb_parse_gmt}
\title{Parse GMT lines}
\usage{
.gsdb_parse_gmt(lines)
}
\arguments{
\item{lines}{Character vector of non-blank GMT lines.}
}
\value{
\verb{list(sets = <named list>, labels = <named chr>)}.
}
\description{
Parse GMT lines
}
\keyword{internal}
