% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdb-file.R
\name{.gsdb_parse_gmx}
\alias{.gsdb_parse_gmx}
\title{Parse GMX lines}
\usage{
.gsdb_parse_gmx(lines)
}
\arguments{
\item{lines}{Character vector of non-blank GMX lines.}
}
\value{
\verb{list(sets = <named list>, labels = <named chr>)}.
}
\description{
Parse GMX lines
}
\keyword{internal}
