% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-db.R
\name{.gsdb_species_dir}
\alias{.gsdb_species_dir}
\title{Species label in directory form}
\usage{
.gsdb_species_dir(species)
}
\arguments{
\item{species}{Character(1).}
}
\value{
Character(1) with spaces replaced by underscores.
}
\description{
Species label in directory form
}
\keyword{internal}
