% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-io.R
\name{.io_first_noncomment_line}
\alias{.io_first_noncomment_line}
\title{First non-\verb{#}-comment line of a text file}
\usage{
.io_first_noncomment_line(path)
}
\arguments{
\item{path}{Path to a text file.}
}
\value{
A single-character line, or \code{character(0)} if the file has none.
}
\description{
First non-\verb{#}-comment line of a text file
}
\keyword{internal}
