% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-db.R
\name{is_gs_db}
\alias{is_gs_db}
\title{Is this a \code{gs_db}?}
\usage{
is_gs_db(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is this a \code{gs_db}?
}
\keyword{internal}
