% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-result.R
\name{new_gs_result}
\alias{new_gs_result}
\title{Low-level \code{gs_result} constructor}
\usage{
new_gs_result(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
A \code{gs_result}.
}
\description{
Attaches the class without checking anything. Use \code{\link[=gs_result]{gs_result()}} unless you
have just validated \code{x} yourself.
}
\keyword{internal}
