% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{\code{x} unless it is \code{NULL}, in which case \code{y}.}
}
\value{
\code{x} or \code{y}.
}
\description{
Default value for NULL
}
\keyword{internal}
