% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatom.R
\name{print.gatom_refs}
\alias{print.gatom_refs}
\title{Print a \code{gatom_refs} object}
\usage{
\method{print}{gatom_refs}(x, ...)
}
\arguments{
\item{x}{A \code{gatom_refs} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{gatom_refs} object
}
\examples{
\dontrun{
print(gatom_refs("Homo sapiens"))
}
}
