% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-matrix.R
\name{print.gs_matrix}
\alias{print.gs_matrix}
\title{Print a \code{gs_matrix}}
\usage{
\method{print}{gs_matrix}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{gs_matrix}.}

\item{n}{Number of rows to show.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{gs_matrix}
}
