% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-db.R
\name{[.gs_db}
\alias{[.gs_db}
\title{Subset a \code{gs_db}, keeping its attributes}
\usage{
\method{[}{gs_db}(x, i)
}
\arguments{
\item{x}{A \code{gs_db}.}

\item{i}{Index (name, position, or logical) of sets to keep.}
}
\value{
A \code{gs_db} with the selected sets.
}
\description{
Subset a \code{gs_db}, keeping its attributes
}
