% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-matrix.R
\name{summary.gs_matrix}
\alias{summary.gs_matrix}
\title{Summarise a \code{gs_matrix}}
\usage{
\method{summary}{gs_matrix}(object, ...)
}
\arguments{
\item{object}{A \code{gs_matrix}.}

\item{...}{Ignored.}
}
\value{
A one-row tibble of dimensions and score range.
}
\description{
Summarise a \code{gs_matrix}
}
