% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-matrix.R
\name{validate_gs_matrix}
\alias{validate_gs_matrix}
\title{Validate a \code{gs_matrix}}
\usage{
validate_gs_matrix(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{x}, invisibly on success; errors otherwise.
}
\description{
Validate a \code{gs_matrix}
}
\keyword{internal}
