% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-result.R
\name{validate_gs_result}
\alias{validate_gs_result}
\title{Validate a \code{gs_result}}
\usage{
validate_gs_result(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{x}, invisibly on success; errors otherwise.
}
\description{
Validate a \code{gs_result}
}
\keyword{internal}
